refactor: Layout 컴포넌트에서 지도에 관한 관심사를 분리한다
